Disable asset fingerprinting for the images/projects/ folder
Project asset paths (ex. <project_name>/tile.jpg) are build dynamically, so ember can't auto-add their fingerprint when building to deploy. Ideally, I would add a build step to somehow manually generate them, but in lieu of that, they are simply disabled for now.

The down side of this is, if any of these assets change (but not their name/path), the old asset may stay cached on devices that have already visited the website for a long time, but this isn't a huge concern for this portfolio site.