Testing Github Pages server routing
This single-page app should handle all routing client-side, but if a user navigates directly to some route, the server still needs to be told to redirect to the root so the client can handle it. It is unclear how to handle this via Github Pages, so a bit of trial-and-error is in order.